Avoid generating planet.xml with broken UTF-8
The input and output of the CSV files was not done consistently using UTF-8
encoding PerlIO filters, so contributors that had non-ASCII names were
mangled in the resulting file.

Autoflush was not enabled on STDEOUT/STDERR, which caused buffering to
happen when running via docker.

If the feed and entry base were not available when extracting entries,
warnings were spewed to STDERR. This has been fixed by not setting entry
base. This could cause some issues with content that rely on base being
set.